requirements.txt
Django==6.0
djangorestframework==3.16.1
django-cors-headers==4.9.0
python-dotenv==1.2.1
djangorestframework_simplejwt==5.5.1
bcrypt==5.0.0
psycopg2-binary==2.9.11
python-telegram-bot==22.5
requests==2.32.5
drf-spectacular==0.29.0
python-dateutil==2.9.0
pytz==2025.2
django-filter==25.2
gunicorn==23.0.0
whitenoise==6.11.0
python-decouple==3.8

